# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Coffee Distribution CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: priority
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: string
        - in: query
          name: channel
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}/addresses:
    get:
      operationId: listCustomerAddresses
      tags: [customers]
      summary: List customer addresses
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerAddress"
    post:
      operationId: createCustomerAddress
      tags: [customers]
      summary: Create a customer address
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerAddressBody"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerAddress"
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /accounting/approvals:
    get:
      operationId: listAccountingApprovals
      tags: [accounting]
      summary: List deliveries awaiting accounting approval
      parameters:
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Approval queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountingApproval"
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, operations, sales, driver, accounting]
        active:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - email
        - role
        - active
        - createdAt
    Customer:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        customerChannel:
          type: string
        segment:
          type: string
        priorityClass:
          type: string
          enum: [A, B, C]
        paymentTerms:
          type: string
        discountLevel:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        businessChannel:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyName
        - contactPerson
        - phone
        - email
        - customerChannel
        - segment
        - priorityClass
        - paymentTerms
        - active
        - businessChannel
        - createdAt
        - updatedAt
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            addresses:
              type: array
              items:
                $ref: "#/components/schemas/CustomerAddress"
          required:
            - addresses
    CustomerAddress:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        addressType:
          type: string
          enum: [billing, delivery]
        label:
          type: ["string", "null"]
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
      required:
        - id
        - customerId
        - addressType
        - street
        - city
        - postalCode
        - country
        - isDefault
    CreateCustomerBody:
      type: object
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        customerChannel:
          type: string
        segment:
          type: string
        priorityClass:
          type: string
          enum: [A, B, C]
        paymentTerms:
          type: string
        discountLevel:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        businessChannel:
          type: string
      required:
        - companyName
        - contactPerson
        - phone
        - email
        - customerChannel
        - segment
        - priorityClass
        - paymentTerms
        - businessChannel
    UpdateCustomerBody:
      type: object
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        customerChannel:
          type: string
        segment:
          type: string
        priorityClass:
          type: string
          enum: [A, B, C]
        paymentTerms:
          type: string
        discountLevel:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        active:
          type: boolean
        businessChannel:
          type: string
    CreateCustomerAddressBody:
      type: object
      properties:
        addressType:
          type: string
          enum: [billing, delivery]
        label:
          type: ["string", "null"]
        street:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        isDefault:
          type: boolean
      required:
        - addressType
        - street
        - city
        - postalCode
        - country
    Lead:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        businessChannel:
          type: string
        businessType:
          type: string
        estimatedMonthlyConsumption:
          type: ["string", "null"]
        preferredCoffeeType:
          type: ["string", "null"]
        requestedMachineType:
          type: ["string", "null"]
        requestedPaymentTerms:
          type: ["string", "null"]
        extraNotes:
          type: ["string", "null"]
        qualificationStatus:
          type: string
          enum: [qualified, manual_review, pending]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - companyName
        - contactPerson
        - phone
        - email
        - businessChannel
        - businessType
        - qualificationStatus
        - createdAt
        - updatedAt
    CreateLeadBody:
      type: object
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        businessChannel:
          type: string
        businessType:
          type: string
        estimatedMonthlyConsumption:
          type: ["string", "null"]
        preferredCoffeeType:
          type: ["string", "null"]
        requestedMachineType:
          type: ["string", "null"]
        requestedPaymentTerms:
          type: ["string", "null"]
        extraNotes:
          type: ["string", "null"]
      required:
        - companyName
        - contactPerson
        - phone
        - email
        - businessChannel
        - businessType
    UpdateLeadBody:
      type: object
      properties:
        qualificationStatus:
          type: string
          enum: [qualified, manual_review, pending]
        extraNotes:
          type: ["string", "null"]
...
    AccountingApproval:
      type: object
      properties:
        id:
          type: integer
        deliveryId:
          type: integer
        customerName:
          type: string
        customerPriority:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        reviewedByName:
          type: ["string", "null"]
        reviewNotes:
          type: ["string", "null"]
        invoiceTriggered:
          type: boolean
        invoiceTriggeredAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - deliveryId
        - customerName
        - customerPriority
        - status
        - invoiceTriggered
        - createdAt
        - updatedAt
...